moved most of the search functions to the navbar except onsubmit and created searchbar component
