+++ client/src/pages/Settings.js
@@ -23,6 +23,15 @@ const useStyles = makeStyles(theme => ({
 }));
 
 const Settings = () => {
+  const classes = useStyles();
+  const history = useHistory();
+  useEffect(() => {
+    if (!localStorage.getItem("email")) handleLogout();
+    axios
+      .get(`/settings/${localStorage.getItem("email")}/company`)
+      .then(res => setCompanyNames(res.data.companies));
+  }, []);
+
   const [companyNames, setCompanyNames] = useState([]);
   const [platforms, setPlatforms] = useState({
     Reddit: true,
@@ -33,54 +42,11 @@ const Settings = () => {
     Shopify: true,
     "Business Insider": true
   });
-  const [searchInput, setSearch] = useState("");
-  const [selectedCompanies, setSelectedCompanies] = useState([]);
-  const [selectedPlatforms, setSelectedPlatforms] = useState([]);
-  const [isPlatformOpen, setPlatformOpen] = useState(false);
-  const [isCompanyOpen, setCompanyOpen] = useState(false);
-
-  const handlePlatformClose = (event) => {
-    setPlatformOpen(false);
-  }
-
-  const handlePlatformOpen = (event) => {
-    setPlatformOpen(true);
-  }
-
-  const handlePlatformChange = (event) => {
-    setSelectedPlatforms(event.target.value);
-  }
-
-  const handleCompanyClose = (event) => {
-    setCompanyOpen(false);
-  }
-
-  const handleCompanyOpen = (event) => {
-    setCompanyOpen(true);
-  }
-
-  const handleCompanyChange = (event) => {
-    setSelectedCompanies(event.target.value);
-  }
 
   const handleSearchSubmit = (event) => {
     event.preventDefault();
     console.log(event.target.searchfield.value, event.target.companyfield.value, event.target.platformfield.value);
-    setSearch("");
   }
-
-  const onSearchChange = (event) => {
-    setSearch(event.target.value);
-  }
-
-  const classes = useStyles();
-  const history = useHistory();
-  useEffect(() => {
-    if (!localStorage.getItem("email")) handleLogout();
-    axios
-      .get(`/settings/${localStorage.getItem("email")}/company`)
-      .then(res => setCompanyNames(res.data.companies));
-  }, []);
 
   const handleLogout = async () => {
     let response = await axios.post("/logout");
@@ -94,20 +60,8 @@ const Settings = () => {
     <div className={classes.dashboardContainer}>
       <Navbar
         showSearch={true}
-        searchInput={searchInput}
-        onSearchChange={onSearchChange}
         platforms={platforms}
-        selectedPlatforms={selectedPlatforms}
-        isPlatformOpen={isPlatformOpen}
-        handlePlatformClose={handlePlatformClose}
-        handlePlatformOpen={handlePlatformOpen}
-        handlePlatformChange={handlePlatformChange}
         companies={companyNames}
-        selectedCompanies={selectedCompanies}
-        isCompanyOpen={isCompanyOpen}
-        handleCompanyClose={handleCompanyClose}
-        handleCompanyOpen={handleCompanyOpen}
-        handleCompanyChange={handleCompanyChange}
         handleSubmit={handleSearchSubmit}
       />
       <Grid container spacing={0}>
